install fix

+++ setup.py
@@ -16,29 +16,35 @@ except ImportError:
 
 home_folder = os.getenv("HOME")
 tmuxer_system_folder = home_folder + "/.tmuxer"
-if not os.path.isdir(tmuxer_system_folder):
-    os.mkdir(tmuxer_system_folder)
 
-if not os.path.isdir(tmuxer_system_folder + '/samples'):
-    os.mkdir(tmuxer_system_folder + '/samples')
+if sys.argv[1] == 'build':
 
-if not os.path.isdir(tmuxer_system_folder + '/tmux_files'):
-    os.mkdir(tmuxer_system_folder + '/tmux_files')
+	if not os.path.isdir(tmuxer_system_folder):
+		os.mkdir(tmuxer_system_folder)
 
-if os.path.exists(tmuxer_system_folder + '/samples/sample.yml'):
-    os.remove(tmuxer_system_folder + '/samples/sample.yml')
+	if not os.path.isdir(tmuxer_system_folder + '/samples'):
+		os.mkdir(tmuxer_system_folder + '/samples')
 
-shutil.copy('samples/sample.yml', tmuxer_system_folder + '/samples/')
+	if not os.path.isdir(tmuxer_system_folder + '/tmux_files'):
+		os.mkdir(tmuxer_system_folder + '/tmux_files')
 
-if os.path.exists(tmuxer_system_folder + '/config'):
-    os.remove(tmuxer_system_folder + '/config')
+	if os.path.exists(tmuxer_system_folder + '/samples/sample.yml'):
+		os.remove(tmuxer_system_folder + '/samples/sample.yml')
 
-config = ConfigParser.RawConfigParser()
-config.add_section('global')
-config.set('global', 'editor', 'vim')
+	shutil.copy('samples/sample.yml', tmuxer_system_folder + '/samples/')
 
-with open(tmuxer_system_folder + '/config', 'w') as configfile:
-    config.write(configfile)
+	if os.path.exists(tmuxer_system_folder + '/config'):
+		os.remove(tmuxer_system_folder + '/config')
+
+	if os.path.exists(tmuxer_system_folder + '/config'):
+		os.remove(tmuxer_system_folder + '/config')
+
+	config = ConfigParser.RawConfigParser()
+	config.add_section('global')
+	config.set('global', 'editor', 'vim')
+
+	with open(tmuxer_system_folder + '/config', 'w') as configfile:
+		config.write(configfile)
 
 distutils.core.setup(name='Tmuxer',
       version='1.0',
